api: complete profile

+++ app/src/main/java/com/bayutb123/tukerin/data/source/remote/response/auth/login/LoginMapper.kt
@@ -10,7 +10,7 @@ fun LoginResponse.toDomain(): User? {
             email = user.email,
             token = user.apiToken,
             isPremium = user.isPremiumUser == "1",
-            rating = user.rating,
+            rating = user.rating.toDouble(),
             trxPoints = user.trxPoints
         )
     }
