+++ app/src/main/java/com/bayutb123/tukerin/data/source/remote/response/user/UserResponseMapper.kt
@@ -8,7 +8,7 @@ fun UserProfileResponse.toModel() : User {
         email = this.user.email,
         token = this.user.apiToken,
         isPremium = this.user.isPremiumUser == "1",
-        rating = this.user.rating,
+        rating = this.user.rating.toDouble(),
         trxPoints = this.user.trxPoints
     )
 }
